Create battle.js
This is basically Codecademy's Dragon Slayer project. I changed it to not be 50/50 chance of dying every turn. I also changed it to roll new damage each turn. I wanted to make this my first Codebit project but I haven't yet learned how to make an HTML page actually use Javascript. The codebit had a script.js tab for JavaScript code but I'll have to learn how to make it work outside of the CodeCademy work space.

I'm honestly really excited to revisit this code and make it more involved. It would be cool to code Pokemon Gen 1 by myself. This is kind of the first step of an RPG battle.